feat:configure resource not find exception for deletar compromissos and deletar compromissos recorrentes

+++ src/main/java/kisiolar/filipe/Viviane/Ai/Compromissos/CompromissosService.java
@@ -126,7 +126,9 @@ public class CompromissosService {
     }
 
     public void deletarCompromissoPorId(long id){
-        //devo reaftorar depois para caso nao exista esse id
+        if(!compromissosRepository.existsById(id)){
+            throw new ResourceNotFindException("Compromisso com ID:" +id +"não foi encontrado");
+        }
         compromissosRepository.deleteById(id);
     }
 }
